requirements.txt
jsonschema>=4.17.0
mkdocs-git-revision-date-localized-plugin==1.5.*
mkdocs-material==9.5.*
pyyaml==6.0.*
rdflib==7.0.0
reportlab==4.2.5
yamllint==1.35.*
